backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.29.0
databases>=0.8.0

# Validation
pydantic>=2.5.3
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
firebase-admin>=6.3.0
bcrypt>=4.1.2

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Machine Learning / Recommendations
scikit-learn>=1.3.2
numpy>=1.26.3

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
